added unet

+++ main.py
@@ -253,7 +253,9 @@ def test(
                     test_start
                 )
             else:
-                extra_image = net.inference((input_data, directions[:21, ...]))
+                extra_image = net.inference(
+                    (input_data, directions[:21, ...].astype(np.float32))
+                )
             image = np.concatenate([lr_image, extra_image])
             log_prediction = np.concatenate([
                 log_b0, log_b0 - bvalues * image
@@ -275,7 +277,9 @@ def test(
                     test_start
                 ) * roi
             else:
-                prediction = net.inference((lr_image, directions[:21, ...]))
+                prediction = net.inference(
+                    (lr_image, directions[:21, ...].astype(np.float32))
+                )
 
         image_nii = nibabel.load(find_file(config['image'], p_path))
         prediction_nii = nibabel.Nifti1Image(
